requirements.txt
Django==2.0.3
inotify==0.2.9
nose==1.3.7
pyaml==17.12.1
pynotify==0.1.1
pyserial==3.4
pytz==2018.3
PyYAML==3.12
redis==2.10.6
